wl-shell-emulator: initial implementation
This modules enables to use wl_shell_surface(s) with ILM APIs.
The surfaceIDs of wl_shell_surface(s) are generated by this module.
The surfaceIDs are starting from a predefined value,
and it increases when a surface is created.

The requests of wl_shell_surface protocol are not yet implemented.

Signed-off-by: Emre Ucan <eucan@de.adit-jv.com>
Signed-off-by: Nobuhiko Tanibata <nobuhiko_tanibata@xddp.denso.co.jp>